api_model/requirements.txt
fastapi
uvicorn
pandas
numpy
python-dateutil
joblib
mlflow 
prefect